% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{lcmc}
\alias{lcmc}
\title{X1 is the input matrix
  X2 is the output matrix
  k is the number of neighbors to consider}
\usage{
lcmc(X1, X2, k)
}
\description{
X1 is the input matrix
  X2 is the output matrix
  k is the number of neighbors to consider
}
