% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lmds-package}
\alias{lmds}
\alias{lmds-package}
\title{Landmark Multi-Dimensional Scalng}
\description{
A fast dimensionality reduction method scaleable to large numbers of samples. Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical Torgerson MDS, but rather than calculating a complete distance matrix between all pairs of samples, only the distances between a set of landmarks and the samples are calculated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dynverse.org/lmds/}
  \item \url{https://github.com/dynverse/lmds}
  \item Report bugs at \url{https://github.com/dynverse/lmds/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@gmail.com} (\href{https://orcid.org/0000-0002-7114-6248}{ORCID}) (zouter)
}

}
