% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{qnx}
\alias{qnx}
\title{X1 is the input matrix
  X2 is the output matrix
  k is the number of neighbors to consider}
\usage{
qnx(X1, X2, ks = NULL, kt = 1)
}
\description{
X1 is the input matrix
  X2 is the output matrix
  k is the number of neighbors to consider
}
