% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{range_kept}
\alias{range_kept}
\title{Since we consistently want to see a spectrum of
how many nearest neighbors are kept, the focus
of this function is to capture 1 to k of the
percent of nearest neighbors kept.}
\usage{
range_kept(X1, X2, k)
}
\description{
Since we consistently want to see a spectrum of
how many nearest neighbors are kept, the focus
of this function is to capture 1 to k of the
percent of nearest neighbors kept.
}
