% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmds.R, R/package.R
\docType{package}
\name{lmds}
\alias{lmds}
\alias{lmds-package}
\title{Landmark MDS}
\usage{
lmds(x, ndim = 3, distance_method = c("euclidean", "pearson",
  "spearman", "cosine", "manhattan"), landmark_method = c("sample"),
  num_landmarks = 500)
}
\arguments{
\item{x}{A matrix, optionally sparse.}

\item{ndim}{The number of dimensions}

\item{distance_method}{The distance metric to use. Options are "euclidean" (default), "pearson", "spearman", "cosine", "manhattan".}

\item{landmark_method}{The landmark selection method to use. Options are "sample" (default).}

\item{num_landmarks}{The number of landmarks to use,}
}
\description{
Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical 'Torgerson' MDS.
LMDS scales to significantly larger data sets since it only computes the distances between
a set of landmarks and all samples. Has built-in support for sparse matrices.

Landmark Multi-Dimensional Scaling (LMDS) is an extension of classical Torgerson MDS.
LMDS aims to produce the same dimensionality reduction as \code{cmdscale()}, but scale
linearly with respect to the number of samples.
}
\examples{
library(Matrix)
x <- Matrix::rsparsematrix(1000, 1000, .01)
lmds(x, ndim = 3)
}
